MVC updated

+++ src/Controller/AddPatientController.java
@@ -47,7 +47,7 @@ public class AddPatientController {
     @FXML
     private void handleSubmit() {
 
-        Patient patient = new Patient(firstname.getText(), lastname.getText(),age.getText(), illness.getText());
+        Patient patient = new Patient(firstname.getText(), lastname.getText(),Integer.valueOf(age.getText()).intValue(), illness.getText());
 
         model.addPatient(patient);
